✅ Rename Galaad extension to Arche : Advanced Rapier Character Handling Engine  ✅ Add a readme for Arche
