+++ src/app/playground/core/extension/arche/types/ArcheAnimationProps.ts
@@ -1,7 +1,7 @@
 import { RigidBodyProps } from "@react-three/rapier";
 import { AnimationSet } from "./AnimationSet";
 
-export type GalaadAnimationProps = {
+export type ArcheAnimationProps = {
   path: string;
   animationSet: AnimationSet;
   children?: React.ReactNode;
