+++ src/app/playground/core/extension/arche/types/CharacterProps.ts
