+++ src/app/playground/core/extension/arche/types/CharacterState.ts
