+++ src/app/playground/core/extension/arche/types/LockCameraProps.ts
